# backend/swagger.yaml
openapi: 3.0.0
info:
  title: M15ASushi - Invoice Creation and Validation API
  description: |
    A comprehensive API for creating UBL 2.4-compliant invoices and validating against PEPPOL standards. 
    
    ## v1 capabilities:
    - Generate UBL 2.4-compliant invoices from JSON objects
    - Validate UBL-XML invoice against Australian and New Zealand PEPPOL standards
    - Secure storage of invoices requiring specific user authentication to access
    - Retrieve invoices in XML format
    
    # Guide to testing:
    
    We strongly recommend using [Postman](https://www.postman.com/) for testing our API as it will allow you to authenticate requests easiest.
    
    1. Copy and paste the following script in **Scripts->Post-response**.
          ```javascript var jsonData = pm.response.json();
          if (jsonData.status === "success" && jsonData.data && jsonData.data.token) {
              pm.environment.set("token", jsonData.data.token);
          }
    2. Register or login using the **/v1/users/register** or **/v1/users/login** endpoints respectively.
    3. Copy the `token`, then go to **Authorization**, select **Bearer Token** for the Auth Type and paste your `token`.
    4. All done! Your token will now be automatically used when you call our routes :)
  version: 1.0.0
servers:
  - url: http://sushi-invoice-application.ap-southeast-2.elasticbeanstalk.com
paths:
...
  /v1/invoices/create:
    post:
      summary: Create a new invoice
      description: |
        This endpoint allows the creation of a new invoice with various fields that are inputted as a JSON object and will be converted to UBL XML format. Each field will be checked for its type and format (see the schema for more details). Required fields are: 
          - invoiceId
          - issueDate
          - buyer
          - supplier
          - total
          - items
      operationId: createInvoice
      tags:
        - Invoices
      parameters:
        - name: token
          in: cookie
          required: true
          schema:
            type: string
      requestBody:
        required: true
        description: Feel free to copy this request body for testing.
        content:
          application/json:
            schema:
              type: object
              properties:
                invoiceId:
                  type: string
                  example: TEST-INVOICE-ID
                supplier:
                  type: string
                  example: supplier name
                buyer:
                  type: string
                  example: buyer name
                total:
                  type: number
                  format: double
                  example: 9.99
                  description: Must be equal the sum of the item cost(s)
                currency:
                  type: string
                  description: Must be a valid 3 letter currency code (ISO 4217)
                  example: "AUD"
                issueDate:
                  type: string
                  format: date
                  description: Must be a past date and be of the format YYYY-MM-DD
                dueDate:
                  type: string
                  format: date
                  description: Must be on or after the issue date and be of the format YYYY-MM-DD
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      name: 
                        type: string
                      count:
                        type: number
                        example: 1
                        description: Must be greater than 0
                      cost:
                        type: number
                        format: double
                        example: 9.99
                buyerAddress:                                    
                  type: object
                  properties:
                    street:
                      type: string
                      description: Street address of the buyer
                    country:
                      type: string
                      description: Country code of the buyer (ISO 3166-1 alpha-2)
                      example: "AU"
                buyerEmail:
                  type: string
                  format: email
                  description: Email address of the buyer
                buyerPhone:
                  type: string
                  description: Phone number of the buyer
                supplierAddress:
                  type: object
                  properties:
                    street:
                      type: string
                      description: Street address of the supplier
                    country:
                      type: string
                      description: Country code of the supplier (ISO 3166-1 alpha-2)
                      example: "AU"
                supplierEmail:
                  type: string
                  format: email
                  description: Email address of the supplier
                supplierPhone:
                  type: string
                  description: Phone number of the supplier
                paymentAccountId:
                  type: string
                  description: Bank account number
                paymentAccountName:
                  type: string
                  description: Account name
                financialInstitutionBranchId:
                  type: string
                  description: Branch identifier such as BSB
      responses:
        "200":
          description: Invoice created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoiceid:
                    type: string
        "400":
          description: If there are any missing requried fields or invalid details provided in the request body. 
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  error: 
                    type: string
                    example: Specific error message
        "401":
          description: Invalid token
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  error: 
                    type: string
                    example: Specific error message
  /v1/invoices/list:  
    get:
      summary: List invoices with optional filters
      description: |
        This endpoint retrieves a list of invoices belonging to the user. 
        It supports optional filters such as limit, offset, sort, and order to customise the results.
  
        **Example route call:**
        `/v1/invoices/list?limit=10&offset=0&sort=issuedate&order=asc`
      operationId: listInvoices
      tags:
        - Invoices
      parameters:
        - name: token
          in: cookie
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: "Number of invoices to return (default: 10)."
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          description: "Pagination: number of invoices to skip (default: 0)."
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: ["issuedate", "duedate", "total"]
          description: "Sort by field"
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum: ["asc", "desc"]
          description: "Sorting order"
      responses:
        "200":
          description: List of invoices retrieved successfully
          content:
            application/json:
              schema: 
                type: object
                properties:
                  data:
                    type: object 
                    properties:
                      count: 
                        type: integer
                        description: Total number of invoices retrieved
                      invoices: 
                        type: array
                        items:
                          type: string
                          example: <Invoice>example</Invoice>
        "400":
          description: |
            Invalid details:
            - limit less than 1
            - offset less than 0
            - invalid sort query
            - invalid order query
        "401":
          description: Invalid token
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  error: 
                    type: string
                    example: Specific error message
...